+++ memberaudit/tests/testdata/esi_client_stub.py
@@ -37,6 +37,12 @@ _endpoints = [
         "character_id",
     ),
     EsiEndpoint(
+        "Character",
+        "get_characters_character_id_roles",
+        "character_id",
+        needs_token=True,
+    ),
+    EsiEndpoint(
         "Contacts",
         "get_characters_character_id_contacts",
         "character_id",